compute auto assigned roles every 5 minutes
